Optimize relationship matching with geographic indexing and fix data type overflow issues
- Fix numeric overflow by upgrading arrays from int16 to int32 across relationship and venue distributors
- Add configurable age range (min_age/max_age) support for romantic relationship eligibility
- Optimize geographic matching with prebuilt MGU/LGU indices for faster lookups
- Replace boolean mask filtering with direct index-based access for improved performance
- Reduce cheating base probability from 30% to 10% for more realistic behavior

+++ may/relationships/friendship_builder.py
@@ -154,8 +154,8 @@ def _process_all_groups_numba(group_starts, group_ends, group_people_flat,
         n_group = len(group_people)
 
         # Get ages and subsets for this group
-        group_ages = np.empty(n_group, dtype=np.int16)
-        group_subsets = np.empty(n_group, dtype=np.int16)
+        group_ages = np.empty(n_group, dtype=np.int32)
+        group_subsets = np.empty(n_group, dtype=np.int32)
 
         for i in range(n_group):
             pid = group_people[i]
@@ -210,7 +210,7 @@ class FriendshipBuilder:
         n_people = len(self.world.population.people)
 
         # Core attributes as contiguous arrays
-        self._ages = np.array([p.age for p in self.world.population.people], dtype=np.int16)
+        self._ages = np.array([p.age for p in self.world.population.people], dtype=np.int32)
         self._n_people = n_people
 
         # Get geographic levels from world.geography
@@ -261,7 +261,7 @@ class FriendshipBuilder:
                         people_by_venue[venue_idx].append(i)
 
         # Dummy subset array (subset filtering not supported with multiple venues)
-        self._person_subset = np.zeros(n_people, dtype=np.int16)
+        self._person_subset = np.zeros(n_people, dtype=np.int32)
 
         # Convert venue data to flattened arrays for Numba
         self._venue_data = self._flatten_groups(people_by_venue)
@@ -392,12 +392,12 @@ class FriendshipBuilder:
             logger.info(f"  Processing source {source_idx}/{total_sources}: {source_name} (pool={pool_type})")
 
             # Parse filters
-            age_range = np.int16(-1)  # -1 means no filter
+            age_range = np.int32(-1)  # -1 means no filter
             require_same_subset = False
 
             for f in filters:
                 if f['attribute'] == 'age' and f['match'] == 'range':
-                    age_range = np.int16(f['range'])
+                    age_range = np.int32(f['range'])
                 elif f['attribute'] == 'subset_name' and f['match'] == 'same':
                     require_same_subset = True
